backend/requirements.txt
flask
flask-cors
tensorflow
numpy
Pillow
pymongo
python-dotenv
